requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
Pillow>=8.3.0
tqdm>=4.62.0
tensorboard>=2.7.0
jupyter>=1.0.0
plotly>=5.3.0
opencv-python>=4.5.0
pandas>=1.3.0